fixed warnings, updated exe

+++ log_entry.cpp
@@ -268,7 +268,7 @@ bool LineToLogEntryIfAny(const std::string& line, LogEntry* e) {
     if (total_healing <= 0) {
       return false;
     }
-    e->hp_diff = total_healing;
+    e->hp_diff = static_cast<float>(total_healing);
   } else if (type == Type::DMG) {
     // 2/26 21:49:12.929  SWING_DAMAGE,Creature-0-4458-469-5988-12017-000056C441,"Broodlord Lashlayer",0xa48,0x0,Player-4476-013D02E7,"Rawrmew-Gehennas",0x10514,0x0,Creature-0-4458-469-5988-12017-000056C441,0000000000000000,84,100,0,0,0,-1,0,0,0,-7612.08,-1096.62,0,5.4551,63,1726,3226,-1,1,0,0,0,nil,nil,1
     // Note: this indexing starts from 0
